• Fix clicking wrong button in a "shop" account #104 • Fix bug when switching to English locale
